gui/requirements.txt
PyQt5>=5.6
sip
requests>=2.0.0
mysql-connector-python
torch
numpy
sounddevice
soundfile
SpeechRecognition
librosa
noisereduce
pocketsphinx